requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.5
newsapi-python==0.2.3
requests==2.17.1
urllib3==1.21.1
websocket-server==0.4
